markdown shortcode Add a paired, async shortcode to render markdown text. Get the project prefix from the projects data file. Remove project.prefix from the initial config.
