+++ src/config/penguinwatch.js
@@ -1,7 +1,6 @@
 module.exports = {
   project: {
     domain: 'penguinwatch.org',
-    name: 'penguin',
-    prefix: 'PZ'
+    name: 'penguin'
   }
 };
