+++ src/config/sciencegossip.js
@@ -1,7 +1,6 @@
 module.exports = {
   project: {
     domain: 'sciencegossip.org',
-    name: 'illustratedlife',
-    prefix: 'SC'
+    name: 'illustratedlife'
   }
 };
